Stream response literals with bounded syntax and 63-bit lengths

+++ imap-codec/src/codec/decode.rs
@@ -48,7 +48,7 @@ pub(crate) struct IMAPParseError<'a, I> {
 pub(crate) enum IMAPErrorKind<'a> {
     Literal {
         tag: Option<Tag<'a>>,
-        length: u32,
+        length: u64,
         mode: LiteralMode,
     },
     BadNumber,
@@ -202,7 +202,7 @@ pub enum CommandDecodeError<'a> {
         tag: Tag<'a>,
 
         /// Literal length.
-        length: u32,
+        length: u64,
 
         /// Literal mode, i.e., sync or non-sync.
         mode: LiteralMode,
@@ -264,7 +264,7 @@ pub enum ResponseDecodeError {
     /// saving the bytes in memory. Or, it can close the connection.
     LiteralFound {
         /// Literal length.
-        length: u32,
+        length: u64,
     },
 
     /// Decoding failed.
